Revert "Bug 1983028 - Part 8: Add VIXL operands and registers. r=spidermonkey-reviewers,iain" for causing bustage  Utils-vixl.h
This reverts commit c15a9ec24e4fd8a72ad01ebcc7a88ba5f5562ddb.

Revert "Bug 1983028 - Part 7: Update VIXL instructions. r=spidermonkey-reviewers,iain"

This reverts commit d6e65f398dec534f6a3d811bcfc316f19dd1be32.

Revert "Bug 1983028 - Part 6: Sync Constants-vixl with upstream. r=spidermonkey-reviewers,iain"

This reverts commit 290a43b49408f649d47ca28474e9ac2b58f9f1a4.

Revert "Bug 1983028 - Part 5: Sync Globals-vixl with upstream. r=spidermonkey-reviewers,iain"

This reverts commit 98b3f9e6e1473e34e319e945c1d398ccde4b21ea.

Revert "Bug 1983028 - Part 4: Sync Cpu-vixl with upstream. r=spidermonkey-reviewers,iain"

This reverts commit 5de49e74de324a0f1a9fbe1bda496ab6961da452.

Revert "Bug 1983028 - Part 3: Sync Cpu-Features-vixl with upstream. r=spidermonkey-reviewers,iain"

This reverts commit 465796412f5a720621d5d9408e67d6ed639d8dc1.

Revert "Bug 1983028 - Part 2: Sync Utils-vixl with upstream. r=spidermonkey-reviewers,iain"

This reverts commit bb30cf385c5ff923609d0f2095d6daf4e0cf4049.

Revert "Bug 1983028 - Part 1: Add ARM64Flags. r=spidermonkey-reviewers,iain"

This reverts commit f9b63375ef55e38d47a6fca2972cc599602706b3.

UltraBlame original commit: aa0732d251e2b54807df4086e4d1c30dafb636f3

+++ js/src/jit/arm64/vixl/Utils-vixl.cpp
@@ -391,7 +391,7 @@ float FPToFloat(double value,
   }
 
   VIXL_UNREACHABLE();
-  return static_cast<float>(value);
+  return value;
 }
